+++ app/bd.sql
@@ -79,11 +79,11 @@ CREATE TABLE `registro` (
   `canal` varchar(100) COLLATE utf8_unicode_ci NOT NULL,
   `inicio` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `fin` datetime DEFAULT NULL,
-  `idReproduccion` int(11) NOT NULL,
+  `idReproduccion` varchar(200) COLLATE utf8_unicode_ci NOT NULL,
   `hostname` varchar(100) COLLATE utf8_unicode_ci NOT NULL,
   `reproductor` varchar(200) COLLATE utf8_unicode_ci NOT NULL,
   `errores` int(11) NOT NULL,
-  `tiempo` int(11) NOT NULL DEFAULT '0',
+  `tiempo` int(50) NOT NULL DEFAULT '0',
   `notificacion_tiempo` tinyint(1) NOT NULL DEFAULT '0'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci ROW_FORMAT=COMPACT;