+++ app/include/menu.php
@@ -5,7 +5,7 @@
 	</button>
 	<div class="collapse navbar-collapse" id="navbarResponsive">
 		<ul class="navbar-nav navbar-sidenav" id="exampleAccordion">
-			<li class="nav-item" data-toggle="tooltip" data-placement="right" title="Dashboard">
+			<li class="nav-item" data-toggle="tooltip" data-placement="right" title="Inicio">
 				<a class="nav-link" href="index.php">
 					<i class="fa fa-fw fa-dashboard"></i>
 					<span class="nav-link-text">Inicio</span>